refactor | move Page component into Book dir
